pyproject.toml
[tool.poetry]
name = "music-web-be"
version = "0.1.0"
description = ""
authors = ["hiep464 <luuhiep2309@gmail.com>"]
readme = "README.md"
packages = [{include = "music_web_be"}]

[tool.poetry.dependencies]
python = "^3.7"
django = "3.2"
pillow = "9.3"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "4.6.0"
django-cors-headers = "3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
